+++ packages/source-kysely/src/utils/create-kysely-mssql-dialect.test.ts
@@ -1,10 +1,10 @@
 import { MssqlDialect } from 'kysely';
 import * as Tedious from 'tedious';
 
-import { createKyselyMssqlDialect } from '../create-kysely-mssql-dialect';
-import { TediousConnUtils } from '../tedious-conn-utils';
+import { createKyselyMssqlDialect } from './create-kysely-mssql-dialect';
+import { TediousConnUtils } from './tedious-conn-utils';
 
-describe('createKyselySqlServerDialect', () => {
+describe('createKyselyMssqlDialect', () => {
   it('should allow to redefine tedious types', () => {
     const jdbcDsn =
       'sqlserver://localhost:1433;database=db;user=sa;password=pwd;trustServerCertificate=true;encrypt=false';
