+++ packages/source-kysely/src/utils/create-kysely-mssql-dialect.ts
@@ -43,12 +43,12 @@ type Params = {
  * @example
  * ```typescript
  * import { default as Tedious } from 'tedious';
- * import { TediousConnUtils, createKyselySqlServerDialect } from '@flowblade/source-kysely';
+ * import { TediousConnUtils, createKyselyMssqlDialect } from '@flowblade/source-kysely';
  *
  * const jdbcDsn = "sqlserver://localhost:1433;database=db;user=sa;password=pwd;trustServerCertificate=true;encrypt=false";
  * const tediousConfig = TediousConnUtils.fromJdbcDsn(jdbcDsn);
  *
- * const dialect = createKyselySqlServerDialect({
+ * const dialect = createKyselyMssqlDialect({
  *  tediousConfig,
  *  // 👉 Optional tarn pool options
  *  poolOptions: {
